Create new Configurator.Utils helper functions module to re-utilize code in forms & controllers
