+++ apps/configurator/lib/configurator_web/controllers/map_configuration_html/index.html.heex
@@ -16,7 +16,7 @@
     <% else %>
       <.button type="button" phx-click={show_modal("initial-positions-#{map_configuration.id}")}>Show</.button>
       <.modal id={"initial-positions-#{map_configuration.id}"}>
-        <p class="whitespace-pre"><%= embed_to_string(map_configuration.initial_positions) %></p>
+        <p class="whitespace-pre"><%= Configurator.Utils.embed_to_string(map_configuration.initial_positions) %></p>
       </.modal>
     <% end %>
   </:col>
@@ -26,7 +26,7 @@
     <% else %>
       <.button type="button" phx-click={show_modal("obstacles-#{map_configuration.id}")}>Show</.button>
       <.modal id={"obstacles-#{map_configuration.id}"}>
-        <p class="whitespace-pre"><%= embed_to_string(map_configuration.obstacles) %></p>
+        <p class="whitespace-pre"><%= Configurator.Utils.embed_to_string(map_configuration.obstacles) %></p>
       </.modal>
     <% end %>
   </:col>
@@ -36,7 +36,7 @@
     <% else %>
       <.button type="button" phx-click={show_modal("bushes-#{map_configuration.id}")}>Show</.button>
       <.modal id={"bushes-#{map_configuration.id}"}>
-        <p class="whitespace-pre"><%= embed_to_string(map_configuration.bushes) %></p>
+        <p class="whitespace-pre"><%= Configurator.Utils.embed_to_string(map_configuration.bushes) %></p>
       </.modal>
     <% end %>
   </:col>
@@ -46,7 +46,7 @@
     <% else %>
       <.button type="button" phx-click={show_modal("pools-#{map_configuration.id}")}>Show</.button>
       <.modal id={"pools-#{map_configuration.id}"}>
-        <p class="whitespace-pre"><%= embed_to_string(map_configuration.pools) %></p>
+        <p class="whitespace-pre"><%= Configurator.Utils.embed_to_string(map_configuration.pools) %></p>
       </.modal>
     <% end %>
   </:col>
@@ -56,7 +56,7 @@
     <% else %>
       <.button type="button" phx-click={show_modal("crates-#{map_configuration.id}")}>Show</.button>
       <.modal id={"crates-#{map_configuration.id}"}>
-        <p class="whitespace-pre"><%= embed_to_string(map_configuration.crates) %></p>
+        <p class="whitespace-pre"><%= Configurator.Utils.embed_to_string(map_configuration.crates) %></p>
       </.modal>
     <% end %>
   </:col>
@@ -67,7 +67,9 @@
       <.button type="button" phx-click={show_modal("square-wall-#{map_configuration.id}")}>Show</.button>
       <.modal id={"square-wall-#{map_configuration.id}"}>
         <p class="whitespace-pre">
-          <%= embed_to_string(map_configuration.square_wall) |> Jason.encode!() |> Jason.Formatter.pretty_print() %>
+          <%= Configurator.Utils.embed_to_string(map_configuration.square_wall)
+          |> Jason.encode!()
+          |> Jason.Formatter.pretty_print() %>
         </p>
       </.modal>
     <% end %>
