+++ apps/configurator/lib/configurator_web/controllers/map_configuration_html/map_configuration_form.html.heex
@@ -11,18 +11,35 @@
     field={f[:initial_positions]}
     type="textarea"
     label="Initial positions"
-    value={embed_to_string(f[:initial_positions].value)}
+    value={Configurator.Utils.embed_to_string(f[:initial_positions].value)}
   />
   <.input
     field={f[:square_wall]}
     type="textarea"
     label="Square Wall"
-    value={embed_to_string(f[:square_wall].value) |> Jason.encode!() |> Jason.Formatter.pretty_print()}
+    value={
+      Configurator.Utils.embed_to_string(f[:square_wall].value) |> Jason.encode!() |> Jason.Formatter.pretty_print()
+    }
   />
-  <.input field={f[:obstacles]} type="textarea" label="Obstacles" value={embed_to_string(f[:obstacles].value)} />
-  <.input field={f[:bushes]} type="textarea" label="Bushes" value={embed_to_string(f[:bushes].value)} />
-  <.input field={f[:pools]} type="textarea" label="Pools" value={embed_to_string(f[:pools].value)} />
-  <.input field={f[:crates]} type="textarea" label="Crates" value={embed_to_string(f[:crates].value)} />
+  <.input
+    field={f[:obstacles]}
+    type="textarea"
+    label="Obstacles"
+    value={Configurator.Utils.embed_to_string(f[:obstacles].value)}
+  />
+  <.input
+    field={f[:bushes]}
+    type="textarea"
+    label="Bushes"
+    value={Configurator.Utils.embed_to_string(f[:bushes].value)}
+  />
+  <.input field={f[:pools]} type="textarea" label="Pools" value={Configurator.Utils.embed_to_string(f[:pools].value)} />
+  <.input
+    field={f[:crates]}
+    type="textarea"
+    label="Crates"
+    value={Configurator.Utils.embed_to_string(f[:crates].value)}
+  />
   <:actions>
     <.button>Save Map configuration</.button>
   </:actions>
