+++ apps/configurator/lib/configurator_web/controllers/version_html/show.html.heex
@@ -22,6 +22,7 @@
   <:item title="Character settings"><.link href={~p"/versions/#{@version}/characters"}>Link</.link></:item>
   <:item title="Game settings"><.link href={~p"/versions/#{@version}/game_configurations"}>Link</.link></:item>
   <:item title="Map settings"><.link href={~p"/versions/#{@version}/map_configurations"}>Link</.link></:item>
+  <:item title="Modes settings"><.link href={~p"/versions/#{@version}/game_mode_configurations"}>Link</.link></:item>
   <:item title="Skill settings"><.link href={~p"/versions/#{@version}/skills"}>Link</.link></:item>
   <:item title="Consumable Items"><.link href={~p"/versions/#{@version}/consumable_items"}>Link</.link></:item>
 </.list>
